Refactor parts of dashboard.py into two new files node_stats.py and util.py
